+++ tests/Method/SendAnimationMethodTest.php
@@ -36,7 +36,7 @@ final class SendAnimationMethodTest extends MethodTestCase
                 'duration' => 100,
                 'width' => 100,
                 'height' => 100,
-                'thumb' => '',
+                'thumbnail' => '',
                 'caption' => 'caption',
                 'caption_entities' => [['type' => 'pre', 'offset' => 0, 'length' => 1]],
                 'parse_mode' => HasParseModeVariableInterface::PARSE_MODE_MARKDOWN,
@@ -45,7 +45,7 @@ final class SendAnimationMethodTest extends MethodTestCase
                 'allow_sending_without_reply' => true,
                 'reply_markup' => self::buildInlineMarkupArray(),
             ],
-            fileMap: ['animation' => true, 'thumb' => true],
+            fileMap: ['animation' => true, 'thumbnail' => true],
             serializableFields: ['reply_markup']
         );
     }